+++ packages/terra-framework-docs/src/terra-dev-site/test/theme-provider/SwitchThemes.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import ThemeProvider from '../../../ThemeProvider';
-import MockThemeComponent from '../common/MockThemeComponent';
+import ThemeProvider from 'terra-theme-provider';
+import MockThemeComponent from './common/MockThemeComponent';
 
 class SwitchThemes extends React.Component {
   constructor(props) {